14-binary_tree_balance.c

+++ binary_trees.h
@@ -37,5 +37,5 @@ size_t binary_tree_size(const binary_tree_t *tree);
 size_t binary_tree_leaves(const binary_tree_t *tree);
 size_t binary_tree_nodes(const binary_tree_t *tree);
 int binary_tree_balance(const binary_tree_t *tree);
-size_t tree_height(const binary_tree_t *tree);
+int tree_height(const binary_tree_t *tree);
 #endif
